requirements.txt
Django==5.2.1
django-filter==24.3
django-widget-tweaks==1.5.0
Pillow==10.4.0